chore: add productPage container
